CardTemplate class moved in header file
